split out platform roles management into a separate module to keep that logic separate (#4226)
