+++ src/platform/platfrom/platform.module.ts
@@ -1,6 +1,5 @@
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
-import { CommunicationModule } from '@domain/communication/communication/communication.module';
 import { LibraryModule } from '@library/library/library.module';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -15,18 +14,14 @@ import { InnovationHubModule } from '@domain/innovation-hub';
 import { KonfigModule } from '@platform/configuration/config/config.module';
 import { MetadataModule } from '@platform/metadata/metadata.module';
 import { StorageAggregatorModule } from '@domain/storage/storage-aggregator/storage.aggregator.module';
-import { AgentModule } from '@domain/agent/agent/agent.module';
-import { NotificationAdapterModule } from '@services/adapters/notification-adapter/notification.adapter.module';
 import { LicensingModule } from '@platform/licensing/licensing.module';
 import { ForumModule } from '@platform/forum/forum.module';
 import { PlatformInvitationModule } from '@platform/invitation/platform.invitation.module';
-import { UserModule } from '@domain/community/user/user.module';
 
 @Module({
   imports: [
     AuthorizationModule,
     AuthorizationPolicyModule,
-    CommunicationModule,
     PlatformAuthorizationPolicyModule,
     LibraryModule,
     ForumModule,
@@ -35,10 +30,7 @@ import { UserModule } from '@domain/community/user/user.module';
     MetadataModule,
     InnovationHubModule,
     LicensingModule,
-    AgentModule,
-    NotificationAdapterModule,
     PlatformInvitationModule,
-    UserModule,
     TypeOrmModule.forFeature([Platform]),
   ],
   providers: [
